# pkg/api/swagger-docs/swagger.yaml
basePath: /
definitions:
  constants.ContainerHealth:
    enum:
    - healthy
    - unhealthy
    - transitioning
    type: string
    x-enum-varnames:
    - ContainerHealthy
    - ContainerUnhealthy
    - ContainerTransitioning
  constants.ContainerState:
    enum:
    - initialized
    - running
    - paused
    - restarting
    - removing
    - stopped
    - dead
    type: string
    x-enum-varnames:
    - ContainerInitialized
    - ContainerRunning
    - ContainerPaused
    - ContainerRestarting
    - ContainerRemoving
    - ContainerStopped
    - ContainerDead
...
  github_com_SENERGY-Platform_mgw-module-lib_model.AuxService:
    properties:
      bind_mounts:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.BindMount'
        description: '{mntPoint:BindMount}'
        type: object
      configs:
        additionalProperties:
          type: string
        description: '{refVar:ref}'
        type: object
      ext_dependencies:
        additionalProperties:
          $ref: '#/definitions/model.ExtDependencyTarget'
        description: '{refVar:ExtDependencyTarget}'
        type: object
      name:
        type: string
      run_config:
        $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.RunConfig'
      srv_references:
        additionalProperties:
          $ref: '#/definitions/model.SrvRefTarget'
        description: '{refVar:SrvRefTarget}'
        type: object
      tmpfs:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.TmpfsMount'
        description: '{mntPoint:TmpfsMount}'
        type: object
      volumes:
        additionalProperties:
          type: string
        description: '{mntPoint:volName}'
        type: object
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.BindMount:
    properties:
      read_only:
        type: boolean
      source:
        type: string
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.ConfigValue:
    properties:
      data_type:
        $ref: '#/definitions/model.DataType'
      default: {}
      delimiter:
        type: string
      is_slice:
        type: boolean
      opt_ext:
        type: boolean
      options: {}
      required:
        type: boolean
      type:
        type: string
      type_opt:
        $ref: '#/definitions/model.ConfigTypeOptions'
    type: object
...
  github_com_SENERGY-Platform_mgw-module-lib_model.HostResource:
    properties:
      required:
        type: boolean
      tags:
        $ref: '#/definitions/model.Set-string'
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.HttpEndpoint:
    properties:
      name:
        type: string
      path:
        description: internal path
        type: string
      port:
        type: integer
      proxy_conf:
        $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.HttpEndpointProxyConf'
      string_sub:
        $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.HttpEndpointStrSub'
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.HttpEndpointProxyConf:
    properties:
      headers:
        additionalProperties:
          type: string
        type: object
      read_timeout:
        $ref: '#/definitions/time.Duration'
      websocket:
        type: boolean
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.HttpEndpointStrSub:
    properties:
      filters:
        additionalProperties:
          type: string
        type: object
      mime_types:
        items:
          type: string
        type: array
      replace_once:
        type: boolean
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.InputGroup:
    properties:
      description:
        type: string
      group:
        type: string
      name:
        type: string
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.Port:
    properties:
      bindings:
        items:
          type: integer
        type: array
      name:
        type: string
      number:
        type: integer
      protocol:
        $ref: '#/definitions/model.PortProtocol'
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.RunConfig:
    properties:
      command:
        items:
          type: string
        type: array
      pseudo_tty:
        type: boolean
      stop_signal:
        type: string
      stop_timeout:
        $ref: '#/definitions/time.Duration'
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.Secret:
    properties:
      required:
        type: boolean
      tags:
        $ref: '#/definitions/model.Set-string'
      type:
        type: string
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.SecretTarget:
    properties:
      item:
        type: string
      ref:
        type: string
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.Service:
    properties:
      bind_mounts:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.BindMount'
        description: '{mntPoint:BindMount}'
        type: object
      configs:
        additionalProperties:
          type: string
        description: '{refVar:ref}'
        type: object
      device_cgroup_rules:
        items:
          type: string
        type: array
      ext_dependencies:
        additionalProperties:
          $ref: '#/definitions/model.ExtDependencyTarget'
        description: '{refVar:ExtDependencyTarget}'
        type: object
      file_groups:
        additionalProperties:
          type: string
        description: '{basePath:ref}'
        type: object
      files:
        additionalProperties:
          type: string
        description: '{mntPoint:ref}'
        type: object
      host_resources:
        additionalProperties:
          $ref: '#/definitions/model.HostResTarget'
        description: '{mntPoint:HostResTarget}'
        type: object
      http_endpoints:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.HttpEndpoint'
        description: '{externalPath:HttpEndpoint}'
        type: object
      image:
        type: string
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.Port'
        type: array
      run_config:
        $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.RunConfig'
      secret_mounts:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.SecretTarget'
        description: '{mntPoint:SecretTarget}'
        type: object
      secret_vars:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.SecretTarget'
        description: '{refVar:SecretTarget}'
        type: object
      srv_references:
        additionalProperties:
          $ref: '#/definitions/model.SrvRefTarget'
        description: '{refVar:SrvRefTarget}'
        type: object
      tmpfs:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.TmpfsMount'
        description: '{mntPoint:TmpfsMount}'
        type: object
      volumes:
        additionalProperties:
          type: string
        description: '{mntPoint:volName}'
        type: object
    type: object
  github_com_SENERGY-Platform_mgw-module-lib_model.TmpfsMount:
    properties:
      mode:
        type: integer
      size:
        type: integer
    type: object
...
  github_com_SENERGY-Platform_mgw-module-manager_lib_models.Deployment:
    properties:
      configs:
        additionalProperties:
          $ref: '#/definitions/models.InterfaceValue'
        type: object
      containers:
        additionalProperties:
          $ref: '#/definitions/models.Container'
        type: object
      created:
        type: string
      enabled:
        type: boolean
      error_msg:
        type: string
      file_groups:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentFileGroup'
        type: object
      files:
        additionalProperties:
          type: string
        description: '{reference:data}'
        type: object
      global_configs:
        additionalProperties:
          type: string
        description: '{reference:globalConfigId}'
        type: object
      has_error:
        type: boolean
      host_resources:
        additionalProperties:
          type: string
        description: '{reference:hostResourceId}'
        type: object
      id:
        type: string
      module_channel:
        type: string
      module_source:
        type: string
      module_version:
        type: string
      secrets:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentSecret'
        type: object
      state:
        description: health state determined by container states
        type: integer
      updated:
        type: string
      volumes:
        additionalProperties:
          type: string
        description: '{reference:name}'
        type: object
    type: object
  github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentFileGroup:
    properties:
      files:
        items:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentFileGroupFile'
        type: array
      id:
        type: string
    type: object
  github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentFileGroupFile:
    properties:
      data:
        type: string
      format:
        type: string
      path:
        type: string
    type: object
  github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentFileGroupUserInput:
    properties:
      data:
        type: string
      format:
        type: string
    type: object
...
  github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentSecret:
    properties:
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/models.DeploymentSecretItem'
        type: array
    type: object
  github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentUserInput:
    properties:
      configs:
        additionalProperties: true
        description: '{ref:value}'
        type: object
      file_groups:
        additionalProperties:
          additionalProperties:
            $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentFileGroupUserInput'
          type: object
        description: '{ref:{path:FileGroupUserInput}}'
        type: object
      files:
        additionalProperties:
          type: string
        description: '{ref:data}'
        type: object
      global_configs:
        additionalProperties:
          type: string
        description: '{ref:configID}'
        type: object
      host_resources:
        additionalProperties:
          type: string
        description: '{ref:resourceID}'
        type: object
      module_id:
        type: string
      secrets:
        additionalProperties:
          type: string
        description: '{ref:secretID}'
        type: object
    type: object
  github_com_SENERGY-Platform_mgw-module-manager_lib_models.Module:
    properties:
      added:
        type: string
      architectures:
        $ref: '#/definitions/model.Set-model_CPUArch'
      author:
        type: string
      aux_img_src:
        $ref: '#/definitions/model.Set-string'
      aux_services:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.AuxService'
        description: '{ref:AuxService}'
        type: object
      channel:
        type: string
      configs:
        allOf:
        - $ref: '#/definitions/model.Configs'
        description: '{ref:ConfigValue}'
      dependencies:
        additionalProperties:
          type: string
        description: '{moduleID:moduleVersion}'
        type: object
      deployment:
        $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.Deployment'
      description:
        type: string
      error_msg:
        type: string
      file_groups:
        allOf:
        - $ref: '#/definitions/model.Set-string'
        description: '{ref}'
      files:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.ModuleFile'
        type: object
      has_error:
        type: boolean
      host_resources:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.HostResource'
        description: '{ref:HostResource}'
        type: object
      id:
        type: string
      inputs:
        $ref: '#/definitions/model.Inputs'
      is_deployed:
        type: boolean
      license:
        type: string
      name:
        type: string
      secrets:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.Secret'
        description: '{ref:Secret}'
        type: object
      services:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.Service'
        description: '{ref:Service}'
        type: object
      source:
        type: string
      tags:
        $ref: '#/definitions/model.Set-string'
      updated:
        type: string
      version:
        type: string
      volumes:
        allOf:
        - $ref: '#/definitions/model.Set-string'
        description: '{volName}'
    type: object
  github_com_SENERGY-Platform_mgw-module-manager_lib_models.ModuleFile:
    properties:
      default_data:
        type: string
      required:
        type: boolean
      source:
        description: optional source path
        type: string
      type:
        type: string
    type: object
  model.ConfigTypeOption:
    properties:
      data_type:
        $ref: '#/definitions/model.DataType'
      value: {}
    type: object
  model.ConfigTypeOptions:
    additionalProperties:
      $ref: '#/definitions/model.ConfigTypeOption'
    type: object
  model.Configs:
    additionalProperties:
      $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.ConfigValue'
    type: object
  model.DataType:
    enum:
    - bool
    - int
    - float
    - string
    type: string
    x-enum-varnames:
    - BoolType
    - Int64Type
    - Float64Type
    - StringType
  model.ExtDependencyTarget:
    properties:
      id:
        type: string
      service:
        type: string
      template:
        type: string
    type: object
  model.HostResTarget:
    properties:
      read_only:
        type: boolean
      ref:
        type: string
    type: object
  model.Input:
    properties:
      description:
        type: string
      group:
        type: string
      name:
        type: string
    type: object
  model.Inputs:
    properties:
      configs:
        additionalProperties:
          $ref: '#/definitions/model.Input'
        description: '{ref:Input}'
        type: object
      file_groups:
        additionalProperties:
          $ref: '#/definitions/model.Input'
        description: '{ref:Input}'
        type: object
      files:
        additionalProperties:
          $ref: '#/definitions/model.Input'
        description: '{ref:Input}'
        type: object
      groups:
        additionalProperties:
          $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-lib_model.InputGroup'
        description: '{ref:InputGroup}'
        type: object
      resources:
        additionalProperties:
          $ref: '#/definitions/model.Input'
        description: '{ref:Input}'
        type: object
      secrets:
        additionalProperties:
          $ref: '#/definitions/model.Input'
        description: '{ref:Input}'
        type: object
    type: object
  model.PortProtocol:
    enum:
    - tcp
    - udp
    type: string
    x-enum-varnames:
    - TcpPort
    - UdpPort
  model.Set-model_CPUArch:
    additionalProperties:
      type: object
    type: object
  model.Set-string:
    additionalProperties:
      type: object
    type: object
  model.SrvRefTarget:
    properties:
      ref:
        type: string
      template:
        type: string
    type: object
...
  models.Container:
    properties:
      alias:
        type: string
      health:
        allOf:
        - $ref: '#/definitions/constants.ContainerHealth'
        description: docker container health
      image_id:
        description: docker image id
        type: string
      name:
        type: string
      state:
        allOf:
        - $ref: '#/definitions/constants.ContainerState'
        description: docker container state
    type: object
...
  models.DeploymentSecretItem:
    properties:
      as_env:
        type: boolean
      as_mount:
        type: boolean
      name:
        type: string
    type: object
...
  models.InterfaceValue:
    properties:
      data_type:
        type: integer
      is_slice:
        type: boolean
      value: {}
    type: object
  models.Job:
    properties:
      description:
        type: string
      end:
        type: string
      id:
        type: string
      start:
        type: string
    type: object
...
  time.Duration:
    enum:
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
info:
  contact: {}
  description: Central service of the MGW ("multi gateway") edge-gateway stack.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: module-manager API
  version: dev
paths:
...
  /deployment-request:
    get:
      description: Get the modules that must be configured to deploy the given modules
        including their dependencies.
      parameters:
      - collectionFormat: csv
        description: filter by module IDs
        in: query
        items:
          type: string
        name: module_ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: modules to be configured
          schema:
            items:
              $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.Module'
            type: array
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
        "503":
          description: error message
          schema:
            type: string
      summary: Get deployment request
      tags:
      - deployments
  /deployments:
    delete:
      description: Delete the deployments of the given modules as a job. All deployments
        are deleted if allow_all is set and no module IDs are given.
      parameters:
      - collectionFormat: csv
        description: filter by module IDs
        in: query
        items:
          type: string
        name: module_ids
        type: array
      - description: allow deletion of all deployments
        in: query
        name: allow_all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: job
          schema:
            $ref: '#/definitions/models.Job'
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
        "503":
          description: error message
          schema:
            type: string
      summary: Delete deployments
      tags:
      - deployments
    post:
      consumes:
      - application/json
      description: Create deployments for the given modules as a job.
      parameters:
      - description: deployment user inputs
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentUserInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: job
          schema:
            $ref: '#/definitions/models.Job'
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
        "503":
          description: error message
          schema:
            type: string
      summary: Create deployments
      tags:
      - deployments
    put:
      consumes:
      - application/json
      description: Update the deployments of the given modules as a job.
      parameters:
      - description: deployment user inputs
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.DeploymentUserInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: job
          schema:
            $ref: '#/definitions/models.Job'
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
        "503":
          description: error message
          schema:
            type: string
      summary: Update deployments
      tags:
      - deployments
  /deployments-disable:
    post:
      consumes:
      - application/json
      description: Disable the deployments of the given modules.
      parameters:
      - description: module IDs
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: IDs of the disabled deployments
          schema:
            items:
              type: string
            type: array
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
        "503":
          description: error message
          schema:
            type: string
      summary: Disable deployments
      tags:
      - deployments
  /deployments-enable:
    post:
      consumes:
      - application/json
      description: Enable the deployments of the given modules.
      parameters:
      - description: module IDs
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: IDs of the enabled deployments
          schema:
            items:
              type: string
            type: array
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
        "503":
          description: error message
          schema:
            type: string
      summary: Enable deployments
      tags:
      - deployments
  /deployments-recreate:
    post:
      consumes:
      - application/json
      description: Recreate the deployments of the given modules as a job.
      parameters:
      - description: module IDs
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: job
          schema:
            $ref: '#/definitions/models.Job'
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
        "503":
          description: error message
          schema:
            type: string
      summary: Recreate deployments
      tags:
      - deployments
...
swagger: "2.0"